batch: Move line sequence comparison into package
The line-matching package owns structural comparison while exact
line-sequence equality remains at the batch package root.

Merge placement relies on both forms of comparison, and their separated
locations make the lower-level matching responsibility harder to follow.

This commit moves exact sequence equality into line_matching while
retaining its selected path for the architecture transition.

The next commit will place exact sequence search beside the equality helper
it uses.
