+++ src/git_stage_batch/batch/line_sequence_equality.py
@@ -1,30 +1,8 @@
-"""Exact byte-line equality helpers for batch operations."""
+"""Compatibility import for line-sequence equality helpers."""
 
-from __future__ import annotations
+import sys as _sys
 
-from collections.abc import Sequence
-
-
-def line_sequences_equal(
-    left: Sequence[bytes],
-    right: Sequence[bytes],
-) -> bool:
-    """Return whether two line sequences contain the same bytes."""
-    return len(left) == len(right) and all(
-        left[index] == right[index]
-        for index in range(len(left))
-    )
+from .line_matching import sequence_equality as _implementation
 
 
-def line_slice_equals(
-    lines: Sequence[bytes],
-    start: int,
-    expected: Sequence[bytes],
-) -> bool:
-    """Return whether a sequence slice equals the expected byte lines."""
-    if start < 0 or start + len(expected) > len(lines):
-        return False
-    return all(
-        lines[start + offset] == expected[offset]
-        for offset in range(len(expected))
-    )
+_sys.modules[__name__] = _implementation
